Added links from the listing page to access the detail page and the delete page.

+++ src/main/java/com/techacademy/CountryService.java
@@ -37,6 +37,9 @@ public class CountryService {
 
     @Transactional
     public void deleteCountry(String code) {
+        if (!repository.findById(code).isPresent()) {
+            throw new IllegalArgumentException("The specified code does not exist: " + code);
+        }
         repository.deleteById(code);
     }
 }
